chore: modify type Integer for Long

+++ src/main/java/ifpb/edu/br/pj/ifpbichos/model/repository/DonatorRepository.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 import ifpb.edu.br.pj.ifpbichos.model.entity.Donator;
 
 @Repository
-public interface DonatorRepository extends JpaRepository<Donator, Integer>{
+public interface DonatorRepository extends JpaRepository<Donator, Long>{
     public boolean existsByLogin(String Login);
 	
 }
